pyproject.toml
[project]
name = "todo-list-app"
version = "0.1.0"
description = ""
authors = [
    {name = "pooriyamm83",email = "mmpooriya83@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "schedule (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "todo_list_app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
